Resolvido conflitos

+++ ecommerce/src/main/java/org/serratec/ecommerce/controllers/EnderecoController.java
@@ -2,7 +2,12 @@ package org.serratec.ecommerce.controllers;
 
 import java.util.List;
 
+import javax.validation.Valid;
+
+import org.serratec.ecommerce.entities.Cliente;
 import org.serratec.ecommerce.entities.Endereco;
+import org.serratec.ecommerce.exceptions.NoSuchElementFoundException;
+import org.serratec.ecommerce.services.ClienteService;
 import org.serratec.ecommerce.services.EnderecoService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -14,6 +19,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -21,23 +27,49 @@ import org.springframework.web.bind.annotation.RestController;
 public class EnderecoController {
 	@Autowired
 	EnderecoService enderecoService;
-	
+
+	@Autowired
+	ClienteService clienteService;
+
 	@GetMapping
 	public ResponseEntity<List<Endereco>> findAllEndereco() {
 		List<Endereco> enderecoList = enderecoService.findAllEndereco();
-		return new ResponseEntity<>(enderecoList, HttpStatus.OK);
+
+		if (enderecoList.isEmpty()) {
+			throw new NoSuchElementFoundException("Nenhum endereço encontrado.");
+		} else {
+			return new ResponseEntity<>(enderecoList, HttpStatus.OK);
+		}
 	}
-	
+
 	@GetMapping("/{id}")
 	public ResponseEntity<Endereco> findEnderecoById(@PathVariable Integer id) {
 		Endereco endereco = enderecoService.findEnderecoById(id);
-		return new ResponseEntity<>(endereco, HttpStatus.OK);
+
+		if (endereco == null) {
+			throw new NoSuchElementFoundException("Não foi possível encontrar o endereço com o id " + id);
+		} else {
+			return new ResponseEntity<>(endereco, HttpStatus.OK);
+		}
 	}
-	
+
 	@PostMapping
-	public ResponseEntity<Endereco> saveEndereco(@RequestBody Endereco endereco) {
+	public ResponseEntity<Endereco> saveEndereco(@Valid @RequestBody Endereco endereco) {
 		return new ResponseEntity<>(enderecoService.saveEndereco(endereco), HttpStatus.CREATED);
 	}
+
+//	@PostMapping("/endereco-completo")
+//	public ResponseEntity<Endereco> salvarEnderecoViaCep(@RequestParam Integer idCliente, @RequestParam String cep,
+//			@RequestParam String numero, @RequestParam String complemento) {
+//		Cliente cliente = clienteService.findClienteById(idCliente);
+//
+//		if (cliente == null) {
+//			throw new NoSuchElementFoundException("Não foi possível encontrar o cliente com o id " + idCliente);
+//		} else {
+//			Endereco endereco = enderecoService.saveEnderecoDTO(cep, numero, complemento);
+//			return new ResponseEntity<>(endereco, HttpStatus.CREATED);
+//		}
+//	}
 
 	@PutMapping
 	public ResponseEntity<Endereco> updateEndereco(@RequestBody Endereco endereco) {
@@ -45,12 +77,16 @@ public class EnderecoController {
 		return new ResponseEntity<>(enderecoAtualizado, HttpStatus.OK);
 	}
 
-
 	@DeleteMapping("/{id}")
 	public ResponseEntity<String> deleteEnderecoById(@PathVariable Integer id) {
-		enderecoService.deleteEnderecoById(id);
-		return new ResponseEntity<>("", HttpStatus.OK);
+		Endereco endereco = enderecoService.findEnderecoById(id);
+		if (endereco == null) {
+			return new ResponseEntity<>("Não foi possível excluir. O Endereco de id = " + id + " não foi encontrado.",
+					HttpStatus.NOT_FOUND);
+		} else {
+			enderecoService.deleteEnderecoById(id);
+			return new ResponseEntity<>("", HttpStatus.OK);
+		}
 	}
 
 }
-
